+++ packages/types/src/__sot-manifest__.ts
@@ -5,7 +5,7 @@ import type {
   MemberServerMetadata,
   MemberWire,
 } from "./encryption.js";
-import type { Member } from "./identity.js";
+import type { Member } from "./entities/member.js";
 
 /**
  * Registry of every domain entity that participates in the types-as-SoT
